Agregado para obtener todas las solicitudes, solo de prueba

+++ src/main/java/ar/edu/unlpam/ing/ProyectoAyDSII/models/DetalleSolicitud.java
@@ -1,12 +1,11 @@
 package ar.edu.unlpam.ing.ProyectoAyDSII.models;
 
-import java.math.BigInteger;
 import lombok.Data;
 
 @Data
 public class DetalleSolicitud {
   private Integer cantidad;
   private String descripcion;
-  private Integer idResiduo;
-  private BigInteger idSolicitud;
+  private Long idResiduo;
+  private Long idSolicitud;
 }
